+++ playwright/compose/keycloak/setup.sh
@@ -48,13 +48,20 @@ kcadm.sh create -r "$TEST_REALM" "client-scopes/$TEST_GROUPS_CLIENT_SCOPE_ID/pro
     -s protocolMapper=oidc-group-membership-mapper \
     -s consentRequired=false \
     -s 'config."claim.name"=groups' \
-    -s 'config."full.path"=false' \
+    -s 'config."full.path"=true' \
     -s 'config."id.token.claim"=true' \
     -s 'config."access.token.claim"=true' \
     -s 'config."userinfo.token.claim"=true'
 
 TEST_GROUP_ID=$(kcadm.sh create -r "$TEST_REALM" groups -s name=Test -i)
+TEST_SUBGROUP1_ID=$(kcadm.sh create -r "$TEST_REALM" "groups/$TEST_GROUP_ID/children" -s name=Group1 -i)
+
 All_GROUP_ID=$(kcadm.sh create -r "$TEST_REALM" groups -s name=All -i)
+All_SUBGROUP1_ID=$(kcadm.sh create -r "$TEST_REALM" "groups/$All_GROUP_ID/children" -s name=Group1 -i)
+All_SUBGROUP2_ID=$(kcadm.sh create -r "$TEST_REALM" "groups/$All_GROUP_ID/children" -s name=Group2 -i)
+
+SUB_GROUP1_ID=$(kcadm.sh create -r "$TEST_REALM" groups -s name=SubGroup1 -i)
+SUB_GROUP2_ID=$(kcadm.sh create -r "$TEST_REALM" groups -s name=SubGroup2 -i)
 
 TEST_CLIENT_ID=$(kcadm.sh create -r "$TEST_REALM" clients -s "name=Warden" -s "clientId=$SSO_CLIENT_ID" -s "secret=$SSO_CLIENT_SECRET" -s "redirectUris=[\"$DOMAIN/*\", \"http://127.0.0.1:$ROCKET_PORT/*\"]" -i)
 
@@ -71,7 +78,7 @@ kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=admin -
 kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=user -s 'description=Admin role'
 
 ## CREATE ORG ROLES
-kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=OrgUnsync -s 'description=Org unsync role'
+kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=OrgNoSync -s 'description=Org unsync role'
 kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=OrgOwner -s 'description=Org owner role'
 kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=OrgAdmin -s 'description=Org admin role'
 kcadm.sh create -r "$TEST_REALM" "clients/$TEST_CLIENT_ID/roles" -s name=OrgManager -s 'description=Org manager role'
@@ -84,23 +91,29 @@ kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER_ID/reset-password" -s type=pa
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER_ID/groups/$TEST_GROUP_ID"
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER_ID/groups/$All_GROUP_ID"
 kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER" --cid "$TEST_CLIENT_ID" --rolename admin
-kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER" --cid "$TEST_CLIENT_ID" --rolename OrgUnsync
+kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER" --cid "$TEST_CLIENT_ID" --rolename OrgNoSync
 
 TEST_USER2_ID=$(kcadm.sh create users -r "$TEST_REALM" -s "username=$TEST_USER2" -s "firstName=$TEST_USER2" -s "lastName=$TEST_USER2" -s "email=$TEST_USER2_MAIL"  -s emailVerified=true -s enabled=true -i)
 kcadm.sh update users/$TEST_USER2_ID/reset-password -r "$TEST_REALM" -s type=password -s "value=$TEST_USER2_PASSWORD" -n
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER2_ID/groups/$TEST_GROUP_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER2_ID/groups/$TEST_SUBGROUP1_ID"
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER2_ID/groups/$All_GROUP_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER2_ID/groups/$All_SUBGROUP1_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER2_ID/groups/$All_SUBGROUP2_ID"
 kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER2" --cid "$TEST_CLIENT_ID" --rolename user
 kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER2" --cid "$TEST_CLIENT_ID" --rolename OrgOwner
 
 TEST_USER3_ID=$(kcadm.sh create users -r "$TEST_REALM" -s "username=$TEST_USER3" -s "firstName=$TEST_USER3" -s "lastName=$TEST_USER3" -s "email=$TEST_USER3_MAIL"  -s emailVerified=true -s enabled=true -i)
 kcadm.sh update users/$TEST_USER3_ID/reset-password -r "$TEST_REALM" -s type=password -s "value=$TEST_USER3_PASSWORD" -n
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER3_ID/groups/$All_GROUP_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER3_ID/groups/$TEST_SUBGROUP1_ID"
 kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER3" --cid "$TEST_CLIENT_ID" --rolename OrgAdmin
 
 TEST_USER4_ID=$(kcadm.sh create users -r "$TEST_REALM" -s "username=$TEST_USER4" -s "firstName=$TEST_USER4" -s "lastName=$TEST_USER4" -s "email=$TEST_USER4_MAIL"  -s emailVerified=true -s enabled=true -i)
 kcadm.sh update users/$TEST_USER4_ID/reset-password -r "$TEST_REALM" -s type=password -s "value=$TEST_USER4_PASSWORD" -n
 kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER4_ID/groups/$All_GROUP_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER4_ID/groups/$SUB_GROUP1_ID"
+kcadm.sh update -r "$TEST_REALM" "users/$TEST_USER4_ID/groups/$SUB_GROUP2_ID"
 kcadm.sh add-roles -r "$TEST_REALM" --uusername "$TEST_USER4" --cid "$TEST_CLIENT_ID" --rolename OrgManager
 
 TEST_USER5_ID=$(kcadm.sh create users -r "$TEST_REALM" -s "username=$TEST_USER5" -s "firstName=$TEST_USER5" -s "lastName=$TEST_USER5" -s "email=$TEST_USER5_MAIL"  -s emailVerified=true -s enabled=true -i)