+++ playwright/global-utils.ts
@@ -199,7 +199,7 @@ export async function startVault(browser: Browser, testInfo: TestInfo, env = {},
 
     console.log(`Starting OIDCWarden`);
     execSync(`docker compose --profile playwright --env-file test.env up -d OIDCWarden`, {
-        env: { LOGIN_RATELIMIT_MAX_BURST: 100, ...env, ...dbConfig(testInfo) },
+        env: { ...env, ...dbConfig(testInfo) },
     });
     await waitFor("/", browser);
     console.log(`OIDCWarden running on: ${process.env.DOMAIN}`);