+++ src/db/models/organization.rs
@@ -1,5 +1,9 @@
+#![allow(unused_qualifications)]
+
 use chrono::{NaiveDateTime, Utc};
 use derive_more::{AsRef, Deref, Display, From};
+use diesel::sql_query;
+use diesel::sql_types::{Nullable, Text};
 use num_traits::FromPrimitive;
 use serde_json::Value;
 use std::{
@@ -15,7 +19,7 @@ use crate::CONFIG;
 use macros::UuidFromParam;
 
 db_object! {
-    #[derive(Identifiable, Queryable, Insertable, AsChangeset)]
+    #[derive(Identifiable, Queryable, QueryableByName, Insertable, AsChangeset)]
     #[diesel(table_name = organizations)]
     #[diesel(treat_none_as_null = true)]
     #[diesel(primary_key(uuid))]
@@ -333,6 +337,18 @@ use crate::db::DbConn;
 use crate::api::EmptyResult;
 use crate::error::MapResult;
 
+#[derive(Debug, QueryableByName)]
+struct OrgGroupSearch {
+    #[diesel(sql_type = Text)]
+    ogs_name_id: String,
+
+    #[diesel(sql_type = Nullable<Text>)]
+    ogs_group: Option<String>,
+
+    #[diesel(sql_type = Nullable<Text>)]
+    ogs_group_uuid: Option<GroupId>,
+}
+
 /// Database methods
 impl Organization {
     pub async fn save(&self, conn: &mut DbConn) -> EmptyResult {
@@ -432,7 +448,7 @@ impl Organization {
         }}
     }
 
-    pub async fn find_main_org_user_email(user_email: &str, conn: &mut DbConn) -> Option<Organization> {
+    pub async fn find_main_org_user_email(user_email: &str, conn: &mut DbConn) -> Option<Self> {
         let lower_mail = user_email.to_lowercase();
 
         db_run! { conn: {
@@ -446,6 +462,40 @@ impl Organization {
                 .first::<OrganizationDb>(conn)
                 .ok().from_db()
         }}
+    }
+
+    pub async fn find_mapped_orgs_and_groups(
+        params: Vec<(String, Option<String>)>,
+        conn: &mut DbConn,
+    ) -> Vec<(String, Option<String>, Self, Option<GroupId>)> {
+        if !params.is_empty() {
+            db_run! { conn: {
+                let values = std::iter::repeat_n("(?, ?)", params.len()).collect::<Vec<&str>>().join(",");
+
+                let mut query = sql_query(format!(r#"
+                    select keys.column1 AS ogs_name_id, keys.column2 AS ogs_group, organizations.*, groups.uuid AS ogs_group_uuid
+                        from ( VALUES {} ) AS keys
+                        left join organizations
+                        left join groups on groups.organizations_uuid = organizations.uuid  AND ( groups.name = keys.column2 or groups.external_id = keys.column1 )
+                        where ( organizations.name = keys.column1 AND groups.name = keys.column2)
+                            OR ((organizations.name = keys.column1 OR organizations.external_id = keys.column1) AND groups.uuid is null AND keys.column2 is null )
+                            OR (groups.external_id = keys.column1 AND keys.column2 is null);
+                "#, values)).into_boxed();
+
+                for (id, group) in params {
+                    query = query.bind::<Text, _>(id).bind::<Nullable<Text>, _>(group);
+                }
+
+                query
+                    .load::<(OrgGroupSearch, OrganizationDb)>(conn)
+                    .expect("Error testing raw query")
+                    .into_iter()
+                    .map(|(ogs, org)| (ogs.ogs_name_id, ogs.ogs_group, org.from_db(), ogs.ogs_group_uuid) )
+                    .collect()
+            }}
+        } else {
+            Vec::new()
+        }
     }
 }